backend/pyproject.toml
[project]
name = "form-reader"
version = "0.1.0"
description = "Web application for automatic reading of test forms."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
dependencies = [
  "opencv-python>=4.9",
  "numpy>=1.26",
  "pymupdf>=1.24",
  "torch>=2.10.0",
  "timm>=1.0.25",
  "torchvision>=0.25.0",
  "fastapi>=0.133.1",
  "uvicorn>=0.41.0",
  "pydantic-settings>=2.13.1",
  "sqlalchemy>=2.0.47",
  "asyncpg>=0.31.0",
  "aioboto3>=15.5.0",
  "alembic>=1.18.4",
  "psycopg2-binary>=2.9.11",
  "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
